// src/shared/enums/FileTypes.ts
enum FileTypes {
	LEVEL = ".level",
	UI_LAYOUT = ".ui",
	COMPONENT = ".component",
	MATERIAL = ".material",
	TEXTURE = ".texture",
	REGISTRY = ".reg",
	PREVIEW = ".preview",
	PRIMITIVE = ".primitive",
	COLLECTION = ".collection",
	PROJECT = ".projection",
	JSON = ".json",
	JAVASCRIPT = ".js",

}

export default FileTypes